Improves handling of TCP session to OVSwitches, corrects certain names
-Added class "StreamHandler" to allow consistent, concurrent management
of the session going to switches. Each switch will have a StreamHandler
and all classes that want to comunicate with the switch should interface
with it.

TODO: abstract the StreamHandler into the switch so that it is more
easily accessible and doesn't need to be passed to each class that wants
to use it