Added a bunch of functionaility. The controller will now remember the member variables of all switches and will restore previously existing OVSwitch objects when a switch re-connects. Logging was also implemented allowing all classes access to the same Java logger via the 'Controller_LOGGER'.
-Added SwitchHandler which will manage the state of all running switches
and allow SwitchSetup to restore previous switches

-Added SwitchSetup which will handle the creation of a OVSwitch object.
When a new connection is detected the SwitchSetup object will be run in
a new thread. It will go through the steps to establish an OpenFlow
connection with the switch and then listen for an OpenFlow Features
Reply message which will let it know which switch it is talking to (via
datapathID). The SwitchSetup object will then query the SwitchHandler to
see if the switch already exists. If it doesnt a new switch is created,
otherwise the existing switch is stopped, its parameters are updated,
and is restarted. In the future this will allow persistent flows to be
easily created and pushed instantly when a switch reconnects.

-Added a Java Logger so that all classes may use a simple logger to
output log messages.