L2 Learning App and Package restructure
-Added functionality for full L2 learning via the API
	-to accomplish this OFMessages had to be converted to byte arrays as
the OFMessage object is non-serializable. Thankfully the OFMessage
object and children support the writeTo(ByteBuffer) method which will
write the message to the given ByteBuffer. Since ByteBuffer is also
non-serializable, the buffer is then converted to a byte array (byte[]).
This is a serializable object and can be sent via RMI to the
application. The reciver of must then parse the byte array back to an
OFMessage object but this is easy as the BasicFactory class has a
parseMessages method.
	
-Packages were slightly restructured so there is now a package solely
for API interfaces.