Added switch link learning, error handling, moved private classes public
All nested private classes *should* be moved to the public space (its
apparently bad design to have them as private unless a reason is needed)

Added some error handling for the stream handler which can now throw
IOException if the stream is closed when trying to send.

The TopologyMapper should now learn switches. A timeout on the
information is needed and triggers should be implemented to clear old
information should new info come in (PortStatus messages informing us a
port goes down).