Hosts are now mapped to an IP and MAC. Corrected some bugs.
-Hosts are now mapped by the HostMapping object. SwitchMappings are a
switch and port number combined with 0 to many HostMappings.
SwitchMappings are also used to indicate just a port on a switch and
thus can also be used with the Link objects. HostMappings have a time to
live specified by the TopologyManager. When a SwitchMapping is created
for the purpose of mapping hosts (therefore not inter-switch links) the
SwitchMapping inherits a ttl of 3 times that of its first Host and every
time a host is either added or updated the ttl is reset to the initial
value. This allows ports without hosts to lose their mapping after a
certain amount of time yet still allow individual HostMappings to
timeout.