Moved to a new application structure
Applications now must bind to the controller interface and register for
an ID. Apps must also provide a requested priority when doing this.

Once registered each application will have an object passed to it to
bind to the actual API for modifying network behavior. Applications will
also have objects created for them on the controller.